+++ src/main/java/ncu/study/autopaper/dao/UserMapper.java
@@ -7,7 +7,7 @@ import ncu.study.autopaper.model.UserKey;
 import org.apache.ibatis.annotations.Param;
 
 public interface UserMapper {
-    int countByExample(UserExample example);
+    long countByExample(UserExample example);
 
     int deleteByExample(UserExample example);
